+++ src/styles/ticket/buyTicket/BuyTicketContentStyled.js
