+++ src/styles/ticket/sellTicket/SellTicketContentStyled.js
